refactor(mapping): add foreign key inputs for relation
