+++ src/strawchemy/graphql/dto.py
@@ -83,9 +83,11 @@ class QueryMetadata:
 
 class StrawchemyDTOAttributes:
     __strawchemy_description__: ClassVar[str] = "GraphQL type"
+    __strawchemy_is_root_aggregation_type__: ClassVar[bool] = False
     __strawchemy_field_map__: ClassVar[dict[DTOKey, GraphQLFieldDefinition[Any, Any]]] = {}
     __strawchemy_query_hook__: QueryHook[Any] | Sequence[QueryHook[Any]] | None = None
-    __strawchemy_is_root_aggregation_type__: ClassVar[bool] = False
+    __strawchemy_filter__: type[Any] | None = None
+    __strawchemy_order_by__: type[Any] | None = None
 
 
 class _Key(Generic[T]):
@@ -475,9 +477,7 @@ class EnumDTO(DTOBase[Any], Enum):
     def field_definition(self) -> GraphQLFieldDefinition[Any, Any]: ...
 
 
-class MappedDataclassGraphQLDTO(StrawchemyDTOAttributes, MappedDataclassDTO[ModelT]):
-    __strawchemy_filter__: type[Any] | None = None
-    __strawchemy_order_by__: type[Any] | None = None
+class MappedDataclassGraphQLDTO(StrawchemyDTOAttributes, MappedDataclassDTO[ModelT]): ...
 
 
 class UnmappedDataclassGraphQLDTO(StrawchemyDTOAttributes, DataclassDTO[ModelT]): ...
