+++ tests/unit/schemas/mutation_and_query.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from strawchemy import Strawchemy
 
 import strawberry
-from tests.unit.models import Fruit
+from tests.unit.models import Color, Fruit
 
 strawchemy = Strawchemy()
 
@@ -16,13 +16,27 @@ class FruitType: ...
 class FruitInput: ...
 
 
+@strawchemy.type(Color, include="all", override=True)
+class ColorType: ...
+
+
+@strawchemy.input(Color, include="all")
+class ColorInput: ...
+
+
 @strawberry.type
 class Query:
     fruit: FruitType = strawchemy.field()
     fruits: list[FruitType] = strawchemy.field()
+
+    color: ColorType = strawchemy.field()
+    colors: list[ColorType] = strawchemy.field()
 
 
 @strawberry.type
 class Mutation:
     create_fruit: FruitType = strawchemy.create_mutation(FruitInput)
     create_fruits: list[FruitType] = strawchemy.create_mutation(FruitInput)
+
+    create_color: ColorType = strawchemy.create_mutation(ColorInput)
+    create_colors: list[ColorType] = strawchemy.create_mutation(ColorInput)
